HIBERNATE-73: extract assertNotSupported to a shared test helper
