+++ src/integrationTest/java/com/mongodb/hibernate/type/UnsupportedTypeAssertions.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2025-present MongoDB, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.mongodb.hibernate.type;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+import com.mongodb.hibernate.internal.FeatureNotSupportedException;
+import org.hibernate.boot.MetadataSources;
+
+public final class UnsupportedTypeAssertions {
+
+    private UnsupportedTypeAssertions() {}
+
+    /**
+     * Asserts that building Hibernate metadata for {@code entityClass} fails at boot because it maps a field of an
+     * unsupported type.
+     */
+    public static void assertNotSupported(Class<?> entityClass) {
+        assertThatThrownBy(() ->
+                        new MetadataSources().addAnnotatedClass(entityClass).buildMetadata())
+                .isInstanceOf(FeatureNotSupportedException.class)
+                .hasMessageMatching(".*persistent attribute .* has .*type .* that is not supported");
+    }
+}
