+++ src/integrationTest/java/com/mongodb/hibernate/type/temporal/SqlTimestampIntegrationTests.java
@@ -16,7 +16,7 @@
 
 package com.mongodb.hibernate.type.temporal;
 
-import static com.mongodb.hibernate.type.temporal.CalendarIntegrationTests.assertNotSupported;
+import static com.mongodb.hibernate.type.UnsupportedTypeAssertions.assertNotSupported;
 import static com.mongodb.hibernate.type.temporal.UnsupportedItems.SqlTimestampItems;
 import static org.junit.jupiter.api.Assertions.assertAll;
